Adds versioned API access.
Modeled after https://www.bignerdranch.com/blog/adding-versions-rails-api/

This provides us a way to verison API access.

Version is supplied in the header:

`curl -H "accept: application/json; version=1" http://127.0.0.1:3000/docs/changes`

By default this is setup to go to version 1 if no version is supplied.
