+++ README.md
@@ -59,6 +59,7 @@ Name | Located In | Description | Required | Schema | Default
 `last_modified` | query | Limit response by an ending datetime| No | datetime in iso8601 | current time
 `page` | query | request a specific page of results | No | integer | 1
 `per_page` | query | Limit the number of results per page | No | integer (1 - 10000) | 100
+`version` | header | Version of the API request eg(`version=1`) | No | integer | 1
 
 ##### Example Response
 ```json
@@ -124,6 +125,7 @@ Name | Located In | Description | Required | Schema | Default
 `last_modified` | query | Limit response by an ending datetime| No | datetime in iso8601 | current time
 `page` | query | request a specific page of results | No | integer | 1
 `per_page` | query | Limit the number of results per page | No | integer (1 - 10000) | 100
+`version` | header | Version of the API request eg(`version=1`) | No | integer | 1
 
 ##### Example Response
 ```json
@@ -169,6 +171,7 @@ The `/collections` endpoint provides a druid list for collections in the public 
 Name | Located In | Description | Required | Schema | Default
 ---- | ---------- | ----------- | -------- | ------ | -------
 `rows` | query | If `0` then only returns the count of collections | No | integer | |
+`version` | header | Version of the API request eg(`version=1`) | No | integer | 1
 
 ##### Example Response
 ```json
