--- app/views/purls/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! 'purls/purl', purl: @purl
