+++ app/views/v1/collections/_collection.json.jbuilder
