+++ app/views/v1/purls/index.json.jbuilder
@@ -1,4 +1,4 @@
 json.purls do
-  json.array! @purls, partial: 'purls/purl', as: :purl
+  json.array! @purls, partial: 'v1/purls/purl', as: :purl
 end
 json.partial! 'shared/paginate', locals: { object: @purls }
