+++ app/views/v1/purls/show.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! 'v1/purls/purl', purl: @purl
